requirements.txt
Flask==1.1.4
Flask-RESTful==0.3.8
prometheus-flask-exporter==0.20.3
PyYAML==5.3.1
requests==2.25.1
# transitory problem https://github.com/pallets/flask/issues/4456
markupsafe==2.0.1
cryptography==39.0.1
toolforge-weld==1.1.0
